api/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
pandas==2.2.2
indic-nlp-library==0.92
requests==2.32.3
gunicorn==22.0.0
langextract==1.0.9
pymilvus==2.3.6
torch==2.2.2
transformers==4.41.2
datasets==2.18.0
accelerate==0.30.1
pytest==8.2.0
pytest-mock==3.14.0
marshmallow==3.21.3
google-generativeai==0.5.4
great_expectations
pandera
tweepy==4.14.0
psycopg2-binary==2.9.9
